add title prop in icon to help with accessibility of non decorative icon
